aggiunto controllo nel service di tavolo

+++ src/main/java/it/prova/pokeronline/service/tavolo/TavoloServiceImpl.java
@@ -12,6 +12,7 @@ import it.prova.pokeronline.model.Tavolo;
 import it.prova.pokeronline.model.Utente;
 import it.prova.pokeronline.repository.tavolo.TavoloRepository;
 import it.prova.pokeronline.service.utente.UtenteService;
+import it.prova.pokeronline.web.api.exception.TavoloConGiocatoriException;
 import it.prova.pokeronline.web.api.exception.TavoloNotFoundException;
 
 @Service
@@ -43,6 +44,14 @@ public class TavoloServiceImpl implements TavoloService {
 	@Override
 	@Transactional
 	public Tavolo aggiorna(Tavolo tavoloInstance) {
+		Tavolo tavoloToUpdate = tavoloRepository.findById(tavoloInstance.getId())
+				.orElseThrow(() -> new TavoloNotFoundException("Tavolo not found con id: " + tavoloInstance.getId()));
+		;
+
+		if (tavoloToUpdate.getGiocatori().size() != 0) {
+			throw new TavoloConGiocatoriException("Non puoi modificare un tavolo se ha dei giocatori associati");
+		}
+
 		return tavoloRepository.save(tavoloInstance);
 	}
 
@@ -62,8 +71,11 @@ public class TavoloServiceImpl implements TavoloService {
 	@Override
 	@Transactional
 	public void rimuovi(Long idToRemove) {
-		tavoloRepository.findById(idToRemove)
+		Tavolo tavoloToRemove = tavoloRepository.findById(idToRemove)
 				.orElseThrow(() -> new TavoloNotFoundException("Tavolo not found con id: " + idToRemove));
+		if (tavoloToRemove.getGiocatori().size() != 0) {
+			throw new TavoloConGiocatoriException("Non puoi eliminare un tavolo se ha dei giocatori associati");
+		}
 		tavoloRepository.deleteById(idToRemove);
 	}
